Switch STO3G water input from Float64 to Float32

+++ example_inputs/sto3g-water.jl
@@ -5,7 +5,7 @@ using InputStructs
 #------------------------------#
 function input_ctrl_flags()
     #Set flags
-    prec::DataType = Float64
+    prec::DataType = Float32
 
     #Form CTRL struct
     CTRL::Ctrl_Flags = Ctrl_Flags(prec)
